[add] add test file and try to implement customer-productor model
