+++ os/test/test.h
@@ -1,0 +1,15 @@
+//
+// Created by 薛英琛 on 2017/4/19.
+//
+
+#ifndef OS_TEST_H
+#define OS_TEST_H
+
+#include <stdio.h>
+#include "../yoroutine.h"
+
+Scheduler * root = Scheduler::getInstance();
+
+
+
+#endif //OS_TEST_H
